[Codes] Fix signature scanner out of range error
An IndexOutOfRangeException occurs if the scanner never finds a result for the input signature because it moves onto the next index in the array and attempts to find the mask one index after, even if it's the last in the list.

Preventing that last iteration fixes this issue and doesn't cause any issues with pre-existing codes, including ones that have more than one signature for different game versions.

This is a necessary fix for codes that share the same signatures and patch the same locations without knowing whether it's been done already.